+++ src/main/java/kr/co/fastcampus/travel/domain/trip/dto/request/TripSaveRequest.java
@@ -1,8 +1,9 @@
-package kr.co.fastcampus.travel.domain.trip.controller.dto;
+package kr.co.fastcampus.travel.domain.trip.dto.request;
 
 import java.time.LocalDate;
 import java.util.List;
 
+import kr.co.fastcampus.travel.domain.itinerary.dto.request.ItinerarySaveRequest;
 import kr.co.fastcampus.travel.domain.trip.entity.Trip;
 import lombok.Builder;
