+++ src/main/java/kr/co/fastcampus/travel/view/InputView.java
@@ -11,8 +11,8 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.function.Predicate;
 import java.util.stream.Collectors;
-import kr.co.fastcampus.travel.domain.trip.controller.dto.ItineraryInfoResponse;
-import kr.co.fastcampus.travel.domain.trip.controller.dto.TripInfoResponse;
+import kr.co.fastcampus.travel.domain.itinerary.dto.response.ItineraryInfoResponse;
+import kr.co.fastcampus.travel.domain.trip.dto.response.TripInfoResponse;
 import kr.co.fastcampus.travel.view.enums.Menu;
 
 public class InputView {
